+++ hoster_debug.go
@@ -1,0 +1,15 @@
+package gohost
+
+import (
+	"net/http"
+)
+
+// serveDebug will start the debug endpoint.
+func (h *Hoster) serveDebug() {
+	// check if debug endpoint is enabled
+	if h.DebugAddr != "" {
+		go func() {
+			http.ListenAndServe(h.DebugAddr, nil)
+		}()
+	}
+}
